Refactor: shared/ui/Index.tsx 분리
