+++ src/pages/PostsManagerPage.tsx
@@ -1,30 +1,13 @@
 import { useEffect, useState } from "react";
 import { Edit2, MessageSquare, Plus, Search, ThumbsDown, ThumbsUp, Trash2 } from "lucide-react";
 import { useLocation, useNavigate } from "react-router-dom";
-import {
-  Button,
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  Input,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-  Textarea,
-} from "../shared/ui";
+import { Input } from "../shared/ui/Input";
+import { Textarea } from "../shared/ui/TextArea";
+import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../shared/ui/Dialog";
+import { Card, CardContent, CardHeader, CardTitle } from "../shared/ui/Card";
+import { Button } from "../shared/ui/Button";
+import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "../shared/ui/Table";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../shared/ui/Select";
 
 const PostsManager = () => {
   const navigate = useNavigate();
